pyproject.toml
[project]
name = "my-comfyui"
version = "0.1.0"
description = "CLI tool to manage ComfyUI workflows and models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "fastapi>=0.115.0",
    "fastmcp>=3.2.4",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
mycomfyui = "my_comfyui.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/my_comfyui"]
